requirements.txt
fastapi
uvicorn
python-telegram-bot==21.5
python-dotenv